+++ src/index.ts
@@ -25,9 +25,10 @@ export { exoticMerkleUpdate } from './boc/cell/exoticMerkleUpdate';
 export { exoticPruned } from './boc/cell/exoticPruned';
 
 // Tuples
-export { Tuple, TupleItem, TupleNull, TupleInt, TupleNaN, TupleCell, TupleBuilder } from './tuple/tuple';
+export { Tuple, TupleItem, TupleItemNull, TupleItemInt, TupleItemNaN, TupleItemCell, TupleItemBuilder } from './tuple/tuple';
 export { parseTuple, serializeTuple } from './tuple/tuple';
 export { TupleReader } from './tuple/reader';
+export { TupleBuilder } from './tuple/builder';
 
 // Types
 export * from './types/_export';