+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## Changede
+- Rename `Tuple*` to `TupleItem` to avoid name clashing
+
+## Added
+- Add `TupleBuilder` helper to build tuples
+
 ## [0.39.0] - 2023-01-06
 
 ## Added